+++ src/api/semaphore.cc
@@ -4,7 +4,7 @@
 
 __BEGIN_SYS
 
-Semaphore::Semaphore(int v): _value(v)
+Semaphore::Semaphore(long v) : _value(v)
 {
     db<Synchronizer>(TRC) << "Semaphore(value=" << _value << ") => " << this << endl;
 }
